pyproject.toml
[project]
name = "20250830-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-google-genai",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-ollama",
    "requests>=2.32.5",
    "python-dotenv>=1.1.1",
    "gradio>=5.46.1",
    "langchain>=0.3.27",
]
